requirements.txt
# Phase 2 - Python Dependencies
# Install with: pip install -r requirements.txt

# Web framework
flask==3.0.0

# Google Gemini API
google-generativeai==0.3.2

# Data validation
pydantic==2.5.0

# Optional but recommended
python-dotenv==1.0.0